Work in progress. First cut.
